Modified level 6
The doors in level 6 now open in such a way as to create a corridor that
the player has to walk through.  This is to restore the original
challenge factor of the level as well as to try to mitigate the problem
of the rat running into a door and unexpectedly reversing direction when
the player tries to jump over it.