Updated license agreement
When I added the "careware clause" message to the license file, I
overlooked the fact that the license text was also present in the source
code files and was output at the start of the program.  This commit
fixes that.

There was a bit of a problem with the message at the start of the game
because I specifically formatted it to fit in an 80x24 terminal.  Maybe
this isn't so much of a problem in real life, but I added a delay which
should fix that.  Maybe that message is redundant and should be removed;
I haven't decided yet.