Added a new enemy, "bat"
A bat enemy has been added to the game.  The only difference between it
and the giant rat is that it moves up and down rather than left and
right.

Because of the way enemies were coded into the game in version 1.0,
bats are currently treated as a special case.  In the future, the code
will have to be kneaded so that flying enemies will be standard; indeed,
much of the enemy code will have to be rewritten to allow for special
behavior.

The bat has been tested and moves, behaves, and collides appropriately.
However, its sprite isn't being made transparent properly for reasons
unknown.  See ``doc/bugs.txt'' for my theory as to why.