Added ability to display level messages
The level generation function has been modified to allow messages to be
displayed when the level starts.  The messages appear at the top of the
screen and appear for a length of time which should reasonably allow the
player to read them.

Messages can be added into the level data by typing them out in-between
two double-quotes.  See the example in ``test-levels''

``doc/levels.txt'' has been updated to explain this new functionality.

In order to display the messages, a font file has been added to the game
files.  The font is Proggy Square, and its license has been copied into
``fonts/readme-proggy.txt''.

The to-do file has been updated.