A system by which the game may be scaled
All of the game's assets are now drawn to a separate "screen" object,
which is scaled up with the display if the SCALE option is set to a
value higher than 1.

Unfortunately this causes a great deal of lag, hence this new branch.
If I can figure out a faster way to do it, this commit will disappear.
If this problem eventually goes away somehow, it will be merged back
into the dev branch.