Changed screen size to 960x540
I was wrong before--it turns out that the `camera' wasn't at all a
factor in whether or not objects are drawn in the scene.  It is now.
This should make it much easier to change the window size.

The game has now been successfully converted to a 16x9 aspect ratio,
causing minimal if any upset of the graphical feel of the game.

A harmless but kind of funny consequence of this is that the borders of
some of the levels are now visible to the player, revealing that they in
fact lead nowhere.

I'm still easily amused.