New tutorial levels
The tutorial levels have been changed (and a new one added) to smooth
the learning curve for the player.  In particular, the player is not
able to die until a few levels later, where the "Press BACKSPACE to
restart" message is not overshadowed by the instruction on how to jump.

Now obviously it's only filthy casuls that will need this, but my mom is
a filthy casul and I think she's pretty great.