Player is now moved while standing on moving walls
It took some fiddling, but I managed to add some special behavior to the
`Mover' class such that if the player is standing ON TOP OF a moving
wall, it will use special behavior to move the player along with it.
THIS SPECIAL BEHAVIOR DOES NOT CHECK FOR COLLISSIONS LIKE THE PLAYER'S
OWN `move' FUNCTION DOES AND MAP DESIGNERS SHOULD BE ADVISED TO USE
MOVING WALLS SPARINGLY.  In theory this shouldn't pose a problem (hence
I haven't added a warning to ``doc/levels.txt'') but one can never be
too careful.

This new behavior for the `Mover' class special-cases itself so that the
behavior is not replecated by children of the `Mover' class such as
`Enemy' (though it's kind of funny watching this bug play out)

I did some testing and the collission detection here seems to work out
okay.  There are a few minor gripes (such as the player not being
crushed to death and the "zap to the top" effect that happens instead)
but also a few amusing quirks (like a little "hop" effect when the
player rides a vertial `Mover' up to the top of its patrol) so I'm
satisfied with the results and have removed the relevant not-a-bug from
``doc/bugs.txt''

It's also worth noting that `Mover's seem to collide pretty well with
other `Wall's and `Door's, just like `Enemy's do, which I didn't even
know, so that's pretty awesome.

Added a small item to the to-do list related to `Mover's: Allow map
designers to choose a specific direction when placing moving walls.

Phew.  This is the longest commit message I've ever written.  I need to
go to bed.