Added ability to choose which direction walls move
I can't be stopped...

In order to allow for greater map design, people creating levels can now
decide on a specific direction for moving platforms to move in when the
map is loaded.

The to-do list and level editing guide have been updated.

The to-do list now specifies choosing specific directions for monsters
to move in rather than moving walls, since that's the next step for
allowing better level design.

I'm DEFINITELY going to bed this time.