Monsters now move in a specific direction
Giant rats and bats can now be placed in the level data such that they
move in a specific direction to synchronize the movements of groups.
In order to make room for the symbols which do this, the in-game
"floating letters" messages have been deprecated.  The code to produce
them still exists in the game, but has been commented out of the level
generation function.

The in-game levels including the "extra levels" have been updated to fit
this change.

Updated relevant documentation.  The changelog also includes changes
which I've forgotten to log.