Re-licensed under the Sunflower Seed License
The license agreement has changed from the MIT license with a careware
clause to the Sunflower Seed License, a highly permissive license based
on the Beer-Ware License, Revision 42.  The copyright notices have been
updated to reflect this.

In order to add a copyright notice to the level file, comment lines have
been implemented, and documentation vis-a-vis level editing have been
updated to reflect this.

Changelog updated, version number advanced.