added introductory levels
I started creating new levels for the game, and then I realized that
these levels which introduce the basic controls and act as an
introduction are really more well-suited to players who are new at the
game.  As such, I've added functionality to have these be optional
levels accessed via the command line for players who need a little
assistance.